+++ tests/phan-interaction/hooksnoop/HookContainer.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace MediaWiki\HookContainer;
+
+class HookContainer {
+	public const NOOP = '*no-op*';
+
+	public function register( string $name, $callback ) {
+	}
+}